requirements.txt
celery<4.2
celery[redis]
django-celery-results<2.0
